makes diagnostics tools generally available
