+++ go/database/mpt/tool/main.go
@@ -12,13 +12,7 @@ package main
 
 import (
 	"fmt"
-	"log"
-	"net/http"
 	"os"
-	"runtime"
-	"runtime/pprof"
-	"runtime/trace"
-	"strings"
 
 	"github.com/urfave/cli/v2"
 )
@@ -75,78 +69,4 @@ func main() {
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)
 	}
-}
-
-func addPerformanceDiagnoses(action cli.ActionFunc) cli.ActionFunc {
-	return func(context *cli.Context) error {
-
-		// Start the diagnostic service if requested.
-		diagnosticPort := context.Int(diagnosticsFlag.Name)
-		startDiagnosticServer(diagnosticPort)
-
-		// Start CPU profiling.
-		cpuProfileFileName := context.String(cpuProfileFlag.Name)
-		if strings.TrimSpace(cpuProfileFileName) != "" {
-			if err := startCpuProfiler(cpuProfileFileName); err != nil {
-				return err
-			}
-			defer stopCpuProfiler()
-		}
-
-		// Start recording a trace.
-		traceFileName := context.String(traceFlag.Name)
-		if strings.TrimSpace(traceFileName) != "" {
-			if err := startTracer(traceFileName); err != nil {
-				return err
-			}
-			defer stopTracer()
-		}
-
-		return action(context)
-	}
-}
-
-func startDiagnosticServer(port int) {
-	if port <= 0 || port >= (1<<16) {
-		return
-	}
-	fmt.Printf("Starting diagnostic server at port http://localhost:%d\n", port)
-	fmt.Printf("(see https://pkg.go.dev/net/http/pprof#hdr-Usage_examples for usage examples)\n")
-	fmt.Printf("Block and mutex sampling rate is set to 100%% for diagnostics, which may impact overall performance\n")
-	go func() {
-		addr := fmt.Sprintf("localhost:%d", port)
-		log.Println(http.ListenAndServe(addr, nil))
-	}()
-	runtime.SetBlockProfileRate(1)
-	runtime.SetMutexProfileFraction(1)
-}
-
-func startCpuProfiler(filename string) error {
-	f, err := os.Create(filename)
-	if err != nil {
-		return fmt.Errorf("could not create CPU profile: %s", err)
-	}
-	if err := pprof.StartCPUProfile(f); err != nil {
-		return fmt.Errorf("could not start CPU profile: %s", err)
-	}
-	return nil
-}
-
-func stopCpuProfiler() {
-	pprof.StopCPUProfile()
-}
-
-func startTracer(filename string) error {
-	traceFile, err := os.Create(filename)
-	if err != nil {
-		return fmt.Errorf("failed to create trace file: %v", err)
-	}
-	if err := trace.Start(traceFile); err != nil {
-		return fmt.Errorf("failed to start trace: %v", err)
-	}
-	return nil
-}
-
-func stopTracer() {
-	trace.Stop()
 }
